diiiq qqa qa! :wq

+++ src/main/java/team/themoment/readygsm/domain/activity/repository/ActivityJpaRepository.java
@@ -5,5 +5,5 @@ import org.springframework.stereotype.Repository;
 import team.themoment.readygsm.domain.activity.entity.ActivityJpaEntity;
 
 @Repository
-public interface ActivityJpaRepository extends JpaRepository<ActivityJpaEntity, Long> {
+public interface ActivityJpaRepository extends JpaRepository<ActivityJpaEntity, Integer> {
 }
